feat: folder_setup tests
